requirements.txt
PySide6>=6.6.0
pywin32>=306
google-genai>=0.1.0
python-dotenv>=1.0.1
